feature 'schedutils' is discared.
The prominent feature of v4.7, a new cpu governor "schedutils", is discarded from our program.

On VMWare (also include other virtual machine tools such as VirtualBox), there is no file for cpu gorvernor, '/sys/devices/system/cpu/cpu*/...'. It is hard to test this feature on our program, so it is discarded.