copy_file_range: discard output of compile and increase copy file size.
In before commit, output of compile the binaries by make is printed on log file.
The useless result is discarded.

And for dramatic difference in test, copying file size is increased.
But the limited memory of guest machine limits it.
Consider default memory usage, we copy 3GB file on 4GB memory system.