madv_free: call sys_madvise() after malloc().
From the original 'ebizzy.c', calling madvise() is added after 'malloc()' (or mmap() but NOT in our test), hint to allocator MADV_FREE or MAVD_DONTNEED.
It can be set by option '-f' to MADV_FREE or to default MADV_DONTNEED.
The main of this commit is followed code lines.

    if (madv_flag)
        syscall(__NR_madvise, p, size, MADV_FREE);
    else
        syscall(__NR_madvise, p, size, MADV_DONTNEED);

Since the glibc 'madvise()' does NOT support MADV_FREE, we call sys_madvise() by syscall() kernel function.
Also, avoid to use glibc malloc(), link to jemalloc library on compile.
It would be adjusted on Makefile and scripts in future work.